+++ frontend/car_rental_app/src/App.js
@@ -3,7 +3,7 @@ import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import LoginForm from './Components/LoginForm/LoginForm';
 import RegisterForm from './Components/RegisterForm/RegisterForm';
-import CarBrowsing from './CarBrowsing';
+import CarBrowsing from './Components/CarBrowsing/CarBrowsing';
 
 function App() {
   return (
