+++ frontend/car_rental_app/src/Components/CarBrowsing/CarCard.jsx
